Moved SafireTestCase base class to make imports easier.
