+++ safire.egg-info/SOURCES.txt
@@ -112,6 +112,12 @@ test/test-data/test-data.emtr.pkl
 test/test-data/test-data.im.ftrs.csv
 test/test-data/test-data.vtext-image-map.csv
 test/test-data/test-data.vtlist
+test/test-data/corpora/test-data.img.icorp
+test/test-data/corpora/test-data.img.mmcorp
+test/test-data/corpora/test-data.img.mmcorp.index
+test/test-data/corpora/test-data.vt.mmcorp
+test/test-data/corpora/test-data.vt.mmcorp.index
+test/test-data/corpora/test-data.vt.vtcorp
 test/test-data/img/0000000323.jpg
 test/test-data/img/0000000326.jpg
 test/test-data/img/0000000714.jpg
