+++ test/test_model_saveload.py
@@ -12,7 +12,8 @@ from safire.learning.models.denoising_autoencoder import DenoisingAutoencoder
 from safire.learning.learners.base_sgd_learner import BaseSGDLearner
 
 import theano
-from test import SafireTestCase
+from test.safire_test_case import SafireTestCase
+
 
 class TestModelSaveLoad(SafireTestCase):
