+++ test/test_multilayer_perceptron.py
@@ -3,7 +3,7 @@ import os
 from safire.data.loaders import MultimodalShardedDatasetLoader, ModelLoader
 from safire.learning.learners import BaseSGDLearner
 from safire.learning.models import MultilayerPerceptron
-from test import SafireTestCase
+from test.safire_test_case import SafireTestCase
 
 __author__ = 'Jan Hajic jr.'
